split util into multiple utils
