+++ packages/twenty-front/src/modules/page-layout/widgets/graph/graphWidgetBarChart/types/FillDateGapsResult.ts
@@ -1,0 +1,6 @@
+import { type GroupByRawResult } from '@/page-layout/widgets/graph/types/GroupByRawResult';
+
+export type FillDateGapsResult = {
+  data: GroupByRawResult[];
+  wasTruncated: boolean;
+};
