+++ packages/twenty-front/src/modules/page-layout/widgets/graph/graphWidgetBarChart/utils/createEmptyDateGroup.ts
@@ -1,0 +1,20 @@
+import { type GroupByRawResult } from '@/page-layout/widgets/graph/types/GroupByRawResult';
+
+export type DimensionValue = string | Date | number | null;
+
+export const createEmptyDateGroup = (
+  dimensionValues: DimensionValue[],
+  keys: string[],
+): GroupByRawResult => {
+  const newItem: GroupByRawResult = {
+    groupByDimensionValues: dimensionValues.map((value) =>
+      value instanceof Date ? value.toISOString() : value,
+    ),
+  };
+
+  for (const key of keys) {
+    newItem[key] = 0;
+  }
+
+  return newItem;
+};
