+++ packages/twenty-front/src/modules/page-layout/widgets/graph/graphWidgetBarChart/utils/fillDateGapsInBarChartData.ts
@@ -1,16 +1,9 @@
 import { BAR_CHART_DATE_GRANULARITIES_WITHOUT_GAP_FILLING } from '@/page-layout/widgets/graph/graphWidgetBarChart/constants/BarChartDateGranularitiesWithoutGapFilling.constant';
-import { generateDateGroupsInRange } from '@/page-layout/widgets/graph/graphWidgetBarChart/utils/generateDateGroupsInRange';
+import { fillDateGapsInOneDimensionalBarChartData } from '@/page-layout/widgets/graph/graphWidgetBarChart/utils/fillDateGapsInOneDimensionalBarChartData';
+import { fillDateGapsInTwoDimensionalBarChartData } from '@/page-layout/widgets/graph/graphWidgetBarChart/utils/fillDateGapsInTwoDimensionalBarChartData';
+import { type SupportedDateGranularity } from '@/page-layout/widgets/graph/graphWidgetBarChart/utils/getDateGroupsFromData';
 import { type GroupByRawResult } from '@/page-layout/widgets/graph/types/GroupByRawResult';
 import { type ObjectRecordGroupByDateGranularity } from 'twenty-shared/types';
-import { isDefined } from 'twenty-shared/utils';
-
-type DimensionValue = string | Date | number | null;
-
-type SupportedDateGranularity =
-  | ObjectRecordGroupByDateGranularity.DAY
-  | ObjectRecordGroupByDateGranularity.MONTH
-  | ObjectRecordGroupByDateGranularity.QUARTER
-  | ObjectRecordGroupByDateGranularity.YEAR;
 
 type FillDateGapsParams = {
   data: GroupByRawResult[];
@@ -19,160 +12,12 @@ type FillDateGapsParams = {
   hasSecondDimension?: boolean;
 };
 
-type FillDateGapsResult = {
-  data: GroupByRawResult[];
-  wasTruncated: boolean;
-};
-
-type OneDimensionalFillParams = {
-  data: GroupByRawResult[];
-  keys: string[];
-  dateGranularity: SupportedDateGranularity;
-};
-
-type TwoDimensionalFillParams = OneDimensionalFillParams;
-
-const createEmptyDateGroup = (
-  dimensionValues: DimensionValue[],
-  keys: string[],
-): GroupByRawResult => {
-  const newItem: GroupByRawResult = {
-    groupByDimensionValues: dimensionValues.map((value) =>
-      value instanceof Date ? value.toISOString() : value,
-    ),
-  };
-
-  for (const key of keys) {
-    newItem[key] = 0;
-  }
-
-  return newItem;
-};
-
-const getDateGroupsFromData = (
-  parsedDates: Date[],
-  dateGranularity: SupportedDateGranularity,
-): { dates: Date[]; wasTruncated: boolean } => {
-  const timestamps = parsedDates.map((date) => date.getTime());
-  const minDate = new Date(Math.min(...timestamps));
-  const maxDate = new Date(Math.max(...timestamps));
-
-  const result = generateDateGroupsInRange({
-    startDate: minDate,
-    endDate: maxDate,
-    granularity: dateGranularity,
-  });
-
-  return { dates: result.dates, wasTruncated: result.wasTruncated };
-};
-
-const fillDateGapsInOneDimensionalBarChartData = ({
-  data,
-  keys,
-  dateGranularity,
-}: OneDimensionalFillParams): FillDateGapsResult => {
-  const existingDateGroupsMap = new Map<string, GroupByRawResult>();
-  const parsedDates: Date[] = [];
-
-  for (const item of data) {
-    const dateValue = item.groupByDimensionValues?.[0];
-
-    if (!isDefined(dateValue)) {
-      continue;
-    }
-
-    const parsedDate = new Date(String(dateValue));
-
-    if (isNaN(parsedDate.getTime())) {
-      continue;
-    }
-
-    parsedDates.push(parsedDate);
-    existingDateGroupsMap.set(parsedDate.toISOString(), item);
-  }
-
-  if (parsedDates.length === 0) {
-    return { data, wasTruncated: false };
-  }
-
-  const { dates: allDates, wasTruncated } = getDateGroupsFromData(
-    parsedDates,
-    dateGranularity,
-  );
-
-  const filledData = allDates.map((date) => {
-    const key = date.toISOString();
-    const existingDateGroup = existingDateGroupsMap.get(key);
-
-    return isDefined(existingDateGroup)
-      ? existingDateGroup
-      : createEmptyDateGroup([date], keys);
-  });
-
-  return { data: filledData, wasTruncated };
-};
-
-const fillDateGapsInTwoDimensionalBarChartData = ({
-  data,
-  keys,
-  dateGranularity,
-}: TwoDimensionalFillParams): FillDateGapsResult => {
-  const existingDateGroupsMap = new Map<string, GroupByRawResult>();
-  const parsedDates: Date[] = [];
-  const uniqueSecondDimensionValues = new Set<DimensionValue>();
-
-  for (const item of data) {
-    const dateValue = item.groupByDimensionValues?.[0];
-
-    if (!isDefined(dateValue)) {
-      continue;
-    }
-
-    const parsedDate = new Date(String(dateValue));
-
-    if (isNaN(parsedDate.getTime())) {
-      continue;
-    }
-
-    parsedDates.push(parsedDate);
-
-    const secondDimensionValue = (item.groupByDimensionValues?.[1] ??
-      null) as DimensionValue;
-    uniqueSecondDimensionValues.add(secondDimensionValue);
-
-    const key = `${parsedDate.toISOString()}_${String(secondDimensionValue)}`;
-    existingDateGroupsMap.set(key, item);
-  }
-
-  if (parsedDates.length === 0) {
-    return { data, wasTruncated: false };
-  }
-
-  const { dates: allDates, wasTruncated } = getDateGroupsFromData(
-    parsedDates,
-    dateGranularity,
-  );
-
-  const filledData = allDates.flatMap((date) =>
-    Array.from(uniqueSecondDimensionValues).map((secondDimensionValue) => {
-      const key = `${date.toISOString()}_${String(secondDimensionValue)}`;
-      const existingDateGroup = existingDateGroupsMap.get(key);
-
-      return isDefined(existingDateGroup)
-        ? existingDateGroup
-        : createEmptyDateGroup([date, secondDimensionValue], keys);
-    }),
-  );
-
-  return { data: filledData, wasTruncated };
-};
-
 export const fillDateGapsInBarChartData = ({
   data,
   keys,
   dateGranularity,
   hasSecondDimension = false,
-}: FillDateGapsParams): FillDateGapsResult => {
+}: FillDateGapsParams): { data: GroupByRawResult[]; wasTruncated: boolean } => {
   if (data.length === 0) {
     return { data, wasTruncated: false };
   }
